CKP/pyproject.toml
[project]
name = "emotion-classification-service"
version = "0.1.0"
description = "FastAPI service using Groq for emotion tagging and celebrate detection (text outputs)."
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.0.0",
    "groq>=0.9.0",
    "httpx>=0.27.0",
    "aiohttp>=3.9.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
test = ["pytest>=8.0.0", "pytest-asyncio>=0.23.0"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

